refactor(folder structure): make components/layout dir
