+++ src/App.vue
@@ -2,13 +2,12 @@
 import { onMounted } from 'vue';
 import { RouterView } from 'vue-router';
 
-// Import components that are visible on all pages of app
-import RouterTitle from '@/components/RouterTitle.vue';
-import Header from '@/components/Header/Header.vue';
-import Footer from '@/components/Footer/Footer.vue';
-import UpdateHandlerDialog from '@/components/UpdateHandlerDialog.vue';
-
+import RouterTitle from '@/components/layout/RouterTitle.vue';
+import Header from '@/components/layout/Header/Header.vue';
+import Footer from '@/components/layout/Footer/Footer.vue';
+import UpdateHandlerDialog from '@/components/layout/UpdateHandlerDialog.vue';
 import { useCustomization } from '@/composables/useCustomization';
+
 const { loadSavedCustomizations } = useCustomization();
 
 onMounted(() => {
